tasks/task 3 - UI Bot/task-3-0-ui-bot/hamagistral/requirements.txt
folium==0.14.0
pandas==1.5.3
requests==2.31.0
streamlit_folium==0.12.0
